feat: avanços no curso

+++ src/pessoas/pessoas.service.ts
@@ -1,19 +1,31 @@
-import { ConflictException, Injectable, NotFoundException } from "@nestjs/common";
+import {
+    ConflictException,
+    ForbiddenException,
+    Injectable,
+    NotFoundException,
+} from "@nestjs/common";
 import { CreatePessoaDTO } from "./dto/create-pessoa.dto";
 import { UpdatePessoaDTO } from "./dto/update-pessoa.dto";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Pessoa } from "./entities/pessoa.entity";
 import { Repository } from "typeorm";
+import { HashingService } from "../auth/hashing/hashing.service";
+import { TokenPayloadDTO } from "../auth/dto/token-payload.dto";
 
 @Injectable()
 export class PessoasService {
-    constructor(@InjectRepository(Pessoa) private readonly pessoaRepository: Repository<Pessoa>) {}
+    constructor(
+        @InjectRepository(Pessoa) private readonly pessoaRepository: Repository<Pessoa>,
+        private readonly hashingService: HashingService,
+    ) {}
 
     async create(dto: CreatePessoaDTO) {
         try {
+            const passwordHash = await this.hashingService.hash(dto.password);
+
             const dadosPessoa = {
                 nome: dto.nome,
-                passwordHash: dto.password,
+                passwordHash,
                 email: dto.email,
             };
 
@@ -41,11 +53,15 @@ export class PessoasService {
         return pessoa;
     }
 
-    async update(id: number, dto: UpdatePessoaDTO) {
+    async update(id: number, dto: UpdatePessoaDTO, tokenPayload: TokenPayloadDTO) {
         const dadosPessoa = {
             nome: dto.nome,
-            passwordHash: dto.password,
         };
+
+        if (dto.password) {
+            const passwordHash = await this.hashingService.hash(dto.password);
+            dadosPessoa["passwordHash"] = passwordHash;
+        }
 
         const pessoa = await this.pessoaRepository.preload({
             id,
@@ -56,11 +72,20 @@ export class PessoasService {
             throw new NotFoundException("Pessoa não encontrada.");
         }
 
+        if (pessoa.id !== tokenPayload.sub) {
+            throw new ForbiddenException("Você não é essa pessoa");
+        }
+
         return this.pessoaRepository.save(pessoa);
     }
 
-    async remove(id: number) {
+    async remove(id: number, tokenPayload: TokenPayloadDTO) {
         const pessoa = await this.findOne(id);
+
+        if (pessoa.id !== tokenPayload.sub) {
+            throw new ForbiddenException("Você não é essa pessoa");
+        }
+
         await this.pessoaRepository.remove(pessoa);
     }
 }
